+++ BusinessObjects/Dtos/Auctions/CreateAuctionRequest.cs
@@ -11,7 +11,7 @@ public class CreateAuctionRequest : IValidatableObject
     public DateTime EndTime { get; set; }
     
     [Range(1,100,ErrorMessage = "Value must be between 1 and 100")]
-    public int StepIncrementPercentage { get; set; }
+    public decimal StepIncrementPercentage { get; set; }
     
 
     [Range(1, int.MaxValue, ErrorMessage = "Error: Minimum deposit must be greater than 0")]
